requirements.txt
Flask==3.0.0
Flask-SQLAlchemy==3.1.1
Flask-Login==0.6.3
Werkzeug==3.0.1
langchain==0.1.0
langchain-openai==0.0.2
openai
python-dotenv==1.0.0
langchain-google-genai